pyproject.toml
[tool.poetry]
name = "git-helper"
version = "0.1.0"
description = ""
authors = ["Artem Poliukhovych <me@2mio.com>"]

[tool.poetry.dependencies]
python = "^3.7"
PyInquirer = "^1.0"
jira = "^2.0"
gitpython = "^3.0"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pylint = "^2.3"
flake8 = "^3.7"
autopep8 = "^1.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
